FEAT: update domain model
Change the GetByID method to take a Long instead of Object since the PK of Movie is a Long.

Dropped DB, removed migrations, and recreated migrations and DB to reflect the above changes.

+++ Migrations/20210321010830_Initial.cs
@@ -10,7 +10,7 @@ namespace MovieCollectionRevisited.Migrations
                 name: "Movies",
                 columns: table => new
                 {
-                    MovieID = table.Column<long>(type: "INTEGER", nullable: false)
+                    MovieID = table.Column<int>(type: "INTEGER", nullable: false)
                         .Annotation("Sqlite:Autoincrement", true),
                     Title = table.Column<string>(type: "TEXT", nullable: false),
                     Year = table.Column<int>(type: "INTEGER", nullable: false),
